2007-07-23 Erik Hjortsberg  <erik.hjortsberg@iteam.se>
	* src/components/ogre/widgets/adapters/atlas/ListAdapter.cpp, src/components/ogre/widgets/adapters/atlas/ListAdapter.h, src/components/ogre/widgets/adapters/atlas/bindings/lua/ListAdapter.pkg: use size_t where suitable

+++ src/components/ogre/widgets/adapters/atlas/ListAdapter.cpp
@@ -45,13 +45,13 @@ ListAdapter::~ListAdapter()
 	removeAdapters();
 }
 
-unsigned int ListAdapter::getSize()
+size_t ListAdapter::getSize()
 {
 	return mAttributes.size();
 }
 
 
-const ::Atlas::Message::Element& ListAdapter::valueOfAttr(unsigned int index) const
+const ::Atlas::Message::Element& ListAdapter::valueOfAttr(size_t index) const
 {
 	static Element emptyElement;
     if (index > mAttributes.size())
